refactor: arrow icon 파일 위치 변경 및 삭제
